Move oauth_requests fixture to test_secure_api.py
