+++ tests/api/test_secure_api.py
@@ -1,5 +1,57 @@
 import json
 
+import pytest
+from connexion.security import SecurityHandlerFactory
+
+
+class FakeResponse:
+    def __init__(self, status_code, text):
+        """
+        :type status_code: int
+        :type text: ste
+        """
+        self.status_code = status_code
+        self.text = text
+        self.ok = status_code == 200
+
+    def json(self):
+        return json.loads(self.text)
+
+
+@pytest.fixture
+def oauth_requests(monkeypatch):
+    class FakeClient:
+        @staticmethod
+        async def get(url, params=None, headers=None, timeout=None):
+            """
+            :type url: str
+            :type params: dict| None
+            """
+            headers = headers or {}
+            if url == "https://oauth.example/token_info":
+                token = headers.get("Authorization", "invalid").split()[-1]
+                if token in ["100", "has_myscope"]:
+                    return FakeResponse(
+                        200, '{"uid": "test-user", "scope": ["myscope"]}'
+                    )
+                if token in ["200", "has_wrongscope"]:
+                    return FakeResponse(
+                        200, '{"uid": "test-user", "scope": ["wrongscope"]}'
+                    )
+                if token == "has_myscope_otherscope":
+                    return FakeResponse(
+                        200, '{"uid": "test-user", "scope": ["myscope", "otherscope"]}'
+                    )
+                if token in ["300", "is_not_invalid"]:
+                    return FakeResponse(404, "")
+                if token == "has_scopes_in_scopes_with_s":
+                    return FakeResponse(
+                        200, '{"uid": "test-user", "scopes": ["myscope", "otherscope"]}'
+                    )
+            return url
+
+    monkeypatch.setattr(SecurityHandlerFactory, "client", FakeClient())
+
 
 def test_security_over_nonexistent_endpoints(oauth_requests, secure_api_app):
     app_client = secure_api_app.test_client()
